Refactor: Update folder and file names as they have been repurposed
