+++ packages/extension/src/serviceWorker/createCookieObject.ts
@@ -25,7 +25,7 @@ import {
 import type { Protocol } from 'devtools-protocol';
 
 /**
- * Create cookie object from cookieStore API cookie object, previously saved parsed cookie object if any, and recently captured request/response cookie header.
+ * Create cookie object from ChromeStorage API cookie object, previously saved parsed cookie object if any, and recently captured request/response cookie header.
  * @param parsedCookie Parsed cookie object from request/response.
  * @param url URL of the cookie from the request/response.
  * @param {Protocol.Network.Cookie[]} cookiesList List cookies from the request.
@@ -112,10 +112,10 @@ export function createCookieObject(
 }
 
 /**
- * Parse cookie attribute values from cookieStore API cookie object, previously saved parsed cookie object if any, and recently captured request/response cookie header.
+ * Parse cookie attribute values from ChromeStorage API cookie object, previously saved parsed cookie object if any, and recently captured request/response cookie header.
  * @param type Cookie attribute type.
  * @param parsedCookieValue Cookie attribute value from the parsed cookie object.
- * @param chromeStoreCookieValue Cookie attribute value from the cookieStore API cookie object.
+ * @param chromeStoreCookieValue Cookie attribute value from the ChromeStorage API cookie object.
  * @param url URL of the cookie from the request/response. (Only required for domain attribute)
  * @param cookieValue cookie value to calculate the size of cookie.
  * @returns {string | boolean | number} Cookie attribute value.
