+++ packages/extension/src/serviceWorker/findPreviousCookieDataObject.ts
@@ -17,7 +17,7 @@
 /**
  * Internal dependencies.
  */
-import type { CookieData } from '../localStore';
+import type { CookieData } from '../store';
 
 /**
  * Find previous cookie object from local storage for given tabId and cookieName.
