+++ packages/extension/src/store/index.ts
@@ -13,6 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default as CookieStore } from './cookieStore';
+export { default as ChromeStorage } from './chromeStorage';
 export { default as PreferenceStore } from './preferenceStore';
 export * from './types';
