+++ packages/extension/src/store/utils/getIndexForAllowListedItem.ts
