+++ packages/extension/src/store/utils/updateCookieBadgeText.ts
