+++ packages/extension/src/store/utils/updateStorage.ts
